Create README.md
This is a project to install Redhat TripleO on in a small Dell Data Centre of 7 x R210II and 2 x R210. For optimisation (and this in itself is a discussion point) all the machines will be VMs on Linux KVM.

I am starting with the basic 2 machine cloud configuration; 1 Compute and 1 Controller (and Ceph). Here is the Openstack deploy script as it stands.

deploy_minimal.sh

THT=/usr/share/openstack-tripleo-heat-templates
openstack overcloud deploy --templates $THT \
-e $THT/environments/network-isolation.yaml \
-e /home/stack/templates/network-environment.yaml \
-e $THT/environments/hyperconverged-ceph.yaml \
-e $THT/environments/ceph-ansible/ceph-ansible.yaml \
--control-flavor oakleyse_tester \
--control-scale 1 \
--compute-flavor oakleyse_tester \
--compute-scale 1 \
--libvirt-type qemu \
--ntp-server pool.ntp.org \
--timeout 180 \
--debug